+++ source/app/service-providers/citeproc/extract-bibtex-attachments.ts
@@ -14,6 +14,7 @@
 
 import path from 'path'
 import { bibtex } from 'astrocite'
+import { BracedComment } from 'astrocite-bibtex'
 import pdfSorter from '@common/util/sort-by-pdf'
 
 const AstrociteAST = bibtex.AST
@@ -35,6 +36,17 @@ export default function extractBibtexAttachments (
   let ast = AstrociteAST.parse(fileContents)
   // Return value will be a fast-access dictionary
   let files = Object.create(null)
+
+  // First we search for the jabref comments containing the files' root directories
+  const comments = ast.children.filter(item => item.kind === 'BracedComment') as BracedComment[]
+  // The format of the value field is 'jabref-meta: fileDirectory*:<path>;'
+  const jabrefComments = comments.filter(item => item.value.startsWith('jabref-meta:'))
+  for (let entry of jabrefComments) {
+    const value = entry.value.split(':').map(e => e.trim())
+    if (value[1].startsWith('fileDirectory')) {
+      baseDir = value[2].replace(/;/g, '')
+    }
+  }
 
   // Now let's see what entries have files attached.
   // Such attributes are stored in properties within the entry.